+++ src/Texture.cpp
@@ -12,13 +12,13 @@ Texture::~Texture()
 	SDL_DestroyTexture(this->texture);
 }
 
-void Texture::Render(const glm::vec2& _position, const int _scale)
+void Texture::Render(const glm::vec2& _position, const double _scale)
 {
 	SDL_Rect rect = SDL_Rect();
 	rect.x = (int)_position.x;
 	rect.y = (int)_position.y;
-	rect.w = this->width * _scale;
-	rect.h = this->height * _scale;
+	rect.w = (int)(this->width * _scale);
+	rect.h = (int)(this->height * _scale);
 
 	SDL_RenderCopy(Environment::GetRenderer(), this->texture, NULL, &rect);
 }
